mv api functions into session class
