+++ qbraid/_warnings.py
@@ -35,13 +35,13 @@ def _check_version():
 
     # pylint: disable=import-outside-toplevel
     from ._version import __version__ as version_local
-    from .api.job_api import _running_in_lab
     from .api.session import QbraidSession
 
-    if not _running_in_lab():
+    session = QbraidSession()
+
+    if not session._running_in_lab():
         return
 
-    session = QbraidSession()
     version_api = session.get("/public/lab/get-sdk-version", params={}).json()
 
     if _warn_new_version(version_local, version_api):
