+++ qbraid/providers/device.py
@@ -22,7 +22,6 @@ from typing import TYPE_CHECKING  # pylint: disable=unused-import
 
 from qbraid import circuit_wrapper
 from qbraid.api import ApiError, QbraidSession
-from qbraid.api.job_api import _qbraid_jobs_enabled, _running_in_lab
 from qbraid.exceptions import QbraidError
 from qbraid.transpiler.exceptions import CircuitConversionError
 
@@ -216,7 +215,7 @@ class QuantumDevice(ABC):
         # qBraid Quantum Jobs proxy is enabled, a document has already been
         # created for this job. So, instead creating a duplicate, we query the
         # user jobs for the `vendorJobId` and return the correspondong `qbraidJobId`.
-        if _running_in_lab() and _qbraid_jobs_enabled(vendor):
+        if session._running_in_lab() and session._qbraid_jobs_enabled(vendor):
             try:
                 job = session.post("/get-user-jobs", json={"vendorJobId": vendor_job_id}).json()[0]
                 return job["qbraidJobId"]
